requirements.txt
feedparser
kafka-python
hdfs
requests
